Add command for automated burnin-in deployments
After this change, processbot will react to a PR comment that contains
the string "bot burnin". It will look for a specific job in Gitlab CI
that matches the most recent commit in the PR and attempt to start the
job. Afterwards it will post the result of this attempt as a comment on
the PR and as a message in the default Matrix channel.